manage s3fs and mysql deps import
